+++ src/scave/octaveexport.cc
@@ -64,10 +64,10 @@ void OctaveExport::close()
 
 std::string OctaveExport::makeUniqueName(const char *nameHint)
 {
-    // first, process the name. Only alphanumeric chars are allowed, the rest
-    // is replaced with underscore
+    // first, process the name. Only alphanumeric chars are allowed, others
+    // are replaced with underscore
     std::string name = nameHint;
-    for (int i=0; i<name.length(); i++)
+    for (int i=0; i<(int)name.length(); i++)
         if (!isalnum(name[i]))
             name[i] = '_';
